add logic for delTask

+++ server/task.routes.js
@@ -1,4 +1,4 @@
-import taskController from "./post.controller.js";
+import taskController from "./task.controller.js";
 export default function (request, response) {
     console.log(request.url);
     console.log(request.method);
@@ -23,9 +23,6 @@ export default function (request, response) {
                 return taskController.getCompleteTasks(request, response) // выводит завершенные задачи
             case '/task/:id':
                 return taskController.getOneTask(request, response) // выводит 1 задачу
-            // response.writeHead(400, { 'Content-Type': 'text/html; charset=utf-8' });
-            // response.end('нет контента')
-            // break;
         }
     } else if (request.method === 'POST') {
         switch (request.url) {
@@ -34,11 +31,9 @@ export default function (request, response) {
                 return taskController.addTask(request, response)
         }
     } else if (request.method === 'DELETE') {
-        switch (request.url) {
-            case '/task/delete':
-                return taskController.deleteTask(request, response)
+        if (request.url.startsWith('/api/delete-task/')) {
+            return taskController.deleteTask(request, response);
         }
-
     } else {
         response.writeHead(400, { 'Content-Type': 'text/html; charset=utf-8' });
         response.end('нет контента')
